TransactionType, ComputationAllowance, Memos, and Signers work (#31)
* WIP more decodable types

* TransactionType, ComputationAllowance, Memos, and Signers work
