+++ xrpl-std/src/core/mod.rs
@@ -5,3 +5,4 @@ pub mod field_codes;
 pub mod ledger_objects;
 pub mod tx;
 pub mod types;
+pub mod locator;
